Stage 1.5b: Convert Overlay controller input to Gecko messages
This change enables scrolling for the Overlay.
In addition to listening to the event from OpenVR, Gecko's WinMouseScrollHandler is updated to scroll based upon a specified widget rather than querying the OS for the HWND/Widget beneath a point on the screen. This allows for scrolling to not be dependent on the window's visible/active state on the desktop.