Stage 2a: Scrolling works on a minimize window
This change fixes an issue where calling ::ScreenToClient returns undefined results for a minimized window. As such, the coordinates for the scroll event were wrong and scrolling would not work.
A new flag, MOUSESCROLL_POINT_IN_WINDOW_COORD, is added to nsIDOMWindowUtils to control when this behavior is used for synthesized scrool events.