Stage 2c: Support focus to FxR window
This change enables OpenVR Overlay Focus events to impact the focus and activation states of its backing Firefox window. With focus, this allows the overlay to invoke the virtual keyboard and send character input to the caret focused in the window.
Note that this also means that the Fx Window for the OpenVR Overlay participates in the same focus management as windows on the desktop, so the overlay can steal focus from a desktop Firefox window and vice-versa.