Stage 3d: Exit a WebVR immersive session with Grip button on controller
This change improves the exitPresent behavior via controller by using the same mechanism as the public DOM function.
Two new members are added to VRControllerState
  exitPresentStartFrameId
  exitPresentStopFrameId
to record this gesture from the VR Process so that it can be checked in the Content process to make the call to VRDisplay.exitPresent.

There are two outstanding issues with this change to be addressed in a subsequent change:
- After exitPresent, there is a 'Waiting...' prompt that appears in the headset since the immersive session is still running without presenting. A later change will end the session after exit presentation.
- Sometimes, a mouse move is needed in the Desktop window to transition presentation multiple times. This is part of a series of issue where nsRefreshDriver and related code have dependencies on a focused or active window. This will be addressed in a later change.