Stage 3e: Allow for play and pause of Video media from controller
This change adds functionality to FxRWindowManager to be able to play and pause media playing in an FxR window. This will be the start of new functionality for transport controls when the overlay is rendering 360 video or theater mode content.
This ability depends on MediaControl Service API, which controls and interacts with all media playing in Firefox.

For now, this is mapped to the 2nd (right) mouse button emulated on the controller.