Stage 3g: Add play/pause functionality to transport controls
This change introduces plumbing for new APIs for chrome JS to interact with FxRWindowManager, allowing for changes to playback state of fullscreen media and for changing projection mode for 360 video.
In this change, the play/pause button now works, and the other buttons simply output to the console to demonstrate the API calls.