Bug: Rendering into VR overlay stops when Desktop window is minimized
This change addresses an issue where composited output is not updated in an FxR window in OpenVR. This behavior makes sense on Desktop where a minimized window is not visible; however, a minimized Desktop window for FxR still needs updated output as the user interacts with the overlay.
To fix this issue, GetClientSize is updated on CompositorWidgetParent to always return the underlying swapchain size for an FxR window.