Bug: Dragging causes an FxR Window to hang
This change fixes an issue where a hang occurs when the controller starts (sometimes, accidentally) a drag operation (i.e., button press with movement before release). During a drag operation, a call is made to ::DoDragDrop, a Windows API that introduces another message pump on the stack. This message pump expects further mouse movements for the drag operation, but since the Gecko messages are generated internally to Firefox from the controller, Windows never gets the mouse move events.
The fix is simply to disable drag and drop for FxR windows.