Stage 7.1: Window Management UI and Window resize
This change updates the FxRPC UI to add buttons for moving and resizing the overlay, and it moves the close button from its temporary position in the navbar.
The functionality for resize is implemented in this change, where the overlay itself is scaled. This is a display/render zoom, rather than a resize that involves a relayout and reflow. A subsequent change will resize the underlying window as well to keep fidelity of the content.
Moving the overlay will be implemented in a subsequent change.