Stage 7.2: Window management UI updates and Window Reposition
This change introduces code that allows a user to move an FxR Overlay Window with the controller from its initial position.
The controllers's input (i.e., mouse input via the controller) moves the overlay left/right (along x-axis) or up/down (along y-axis). To move the overlay forward/backward (along z-axis), the user must physically move (i.e., change the position of the HMD).

The core of this change is FxRWindowManager::HandleOverlayMove, which contains plenty of documentation for the calculations involved.

This change includes other fixes:
- Since the overlay is always repositioned via a center point, a modal icon shows up when moving the overlay to indicate that move is occuring
- A new ChromeUtils API is introduced so that Chrome JS can initiate window move
- The default width of the overlay is reduced
- Window resize UI now appears from the bottom rather than replaces the navigation bar

Shoutout to Patrick prototype for this behavior
https://github.com/MozillaReality/gecko-dev/commit/003eae3be0ebcb547ccb96a2e90de31d2dca3b96