Bug: Update Home Page and navbar behavior
This change updates the URL to the home page to allow for interactive content to be skipped without checking WebVR. This allows for the WebXR content feed to load without enumerating VR displays, which will cause an existing immersive scene to exit.
Further, the navbar will no longer show the homepage URL and has default text similarly to Desktop Firefox and FxR on Android.