Stage 7.3: Window Interactivity with Immersive Scenes
This change adds a way for the user to control when to relinquish control to an Immersive App from the Overlay.
When an overlay is running, it takes over controller input such that the immersive scene will not get controller input. Thus, there needs to be a way to toggle input between the overlay and the immersive scene.
To relinquish control, a button is added to the window management bar at the top of the overlay.
To regain control, there are two options added with this change:
1) From the Desktop, click the same button again to regain control
2) Start FxR again, which will detect that an instance is already running and will resume interactivity. This option is to support controlling from the headset (by launching the app again) and needs to be tested.

The basis of this change is modifying the flag
    VROverlayFlags_MakeOverlaysInteractiveIfVisible
under the conditions explained above.