BUG: Controller doesn't work for WebVR/XR launched from FxR
This change fixes an issue where controllers don't appear/work when launched from an FxR window.
It seems that the root cause is that, though these are two different processes (main and VR) with two different applications (Overlay and Scene), they still share the same appkey, which ties some of configuration together. Specifically, controller actions and bindings are shared.
OpenVR requires input to be setup before it is used; since the overlay uses the input first, WebVR's input setup doesn't apply, which caused
  VRInputError_InvalidHandle
to be returned from functions that would get input state.

This change, thus, has the controller setup happen in the main process when FxR is initialized, which is reused when the VR process is setup.