BUG: Unable to select inside text of navbar
This change allows the user to set the cursor inside of text in the navbar. This brings parity with a conventional navbar:
- First click highlights all of the text
- Subsequent click places the caret at the clicked location in the text
Further, Firefox no longer changes focus when the virtual keyboard stays focus so that the overlay window appears to still have focus when typing onto the virtual keyboard.