PoC: Support for Multi-GPU machines
For machines with 2+ GPU, this change ensures that Firefox runs on the same DXGI adapter as OpenVR.
Upon creating an adapter during Firefox startup, DeviceManagerDx::GetDXGIAdapter now queries FxRWindowManager for which adapater to use. FxRWindowManager determines this based on what the VR platform reports.
This reported adapter index is stored in gfxvars so that it can be propogated and accessed to other processes, including the GPU process.

This is an update to a related change
  https://github.com/MozillaReality/gecko-dev/commit/e121c75b2cc85810d87cae196588ed11b5ab42bd
which was based upon doing a device reset to another adapter. Bug 1656034 changes how Angle how devices are created, which conflicts with this original change.
This updated change suffices for the embedded FxR PC case, but this change will need to be revisited if targeted for mozilla-central.